+++ tests/lang/glsl.glsl
@@ -1,0 +1,10 @@
+// 10 lines 5 code 3 comments 2 blanks
+#version 330 core
+
+in vec3 fragPos;
+/* Lighting uniforms */
+uniform vec3 lightColor;
+uniform vec3 objectColor;
+
+// Output
+out vec4 FragColor;