+++ tests/lang/objc.m
@@ -1,0 +1,10 @@
+// 10 lines 6 code 3 comments 1 blank
+#import <Foundation/Foundation.h>
+
+/* Main entry point */
+int main(int argc, const char *argv[]) {
+    // Create pool
+    @autoreleasepool {
+        NSLog(@"Hello");
+    }
+}