+++ tests/lang/pascal.pas
@@ -1,0 +1,10 @@
+// 10 lines 5 code 4 comments 1 blank
+program Hello;
+
+{ Main program }
+uses SysUtils;
+(* Nested { comment } here *)
+// Entry point
+begin
+  WriteLn('Hello, World!');
+end.