+++ tests/lang/perl.pl
@@ -1,0 +1,11 @@
+# 11 lines 3 code 6 comments 2 blanks
+#!/usr/bin/env perl
+use strict;
+
+# Main program
+my $name = "World";
+print "Hello, $name!\n";
+
+=pod
+This is documentation.
+=cut